App now runs as webapp
Reconfigured docker resources to run with nginx as a reverseproxy and
the python app to run on 0.0.0.0:5000. Also made sure the configuration
is such that the app can run seamlessly on AWS EC2 instances.

Added a basic controller and a route to act as the API webhook for
pagerduty to send alerts to